requirements_working.txt
certifi==2026.1.4
future==1.0.0
geopandas==1.1.2
iso8601==2.1.0
numpy==2.4.2
packaging==26.0
pandas==3.0.0
pyogrio==0.12.1
pyproj==3.7.2
pyrtlsdr==0.3.0
pyserial==3.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
serial==0.0.97
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
wxPython==4.2.5
